Make tasker work in Docker and make log streaming optional
