+++ components/fetch-code/habitat/plan.sh
@@ -7,4 +7,5 @@ pkg_upstream_url="https://github.com/chuckleheads/hurtlocker"
 pkg_license=('Apache-2.0')
 pkg_maintainer="The Habitat Maintainers <humans@habitat.sh>"
 pkg_bin_dirs=(bin)
+pkg_deps=(core/cacerts)
 pkg_build_deps=(core/go core/git core/dep)
