+++ components/tasker/pkg/promote/promote.go
@@ -1,6 +1,10 @@
 package promote
 
 import (
+	"context"
+	"log"
+	"time"
+
 	pb "github.com/chuckleheads/hurtlocker/components/logsrv/logrecv"
 	"github.com/chuckleheads/hurtlocker/components/tasker/pkg/cmd"
 	"github.com/spf13/viper"
@@ -10,10 +14,19 @@ type PromoteCli struct {
 	logsrv pb.LogRecv_ReceiveLogsClient
 }
 
-func New(logsrv pb.LogRecv_ReceiveLogsClient) PromoteCli {
-	return PromoteCli{
-		logsrv,
+func New() PromoteCli {
+	promoteCli := PromoteCli{}
+	if viper.GetBool("enable-log-stream") {
+		client := cmd.LogSrvClient()
+		ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
+		defer cancel()
+		stream, err := client.ReceiveLogs(ctx)
+		if err != nil {
+			log.Fatalf("%v.RecordRoute(_) = _, %v", client, err)
+		}
+		promoteCli.logsrv = stream
 	}
+	return promoteCli
 }
 
 func (b *PromoteCli) Promote() {
