+++ components/tasker/sample_config/build.sample.yaml
@@ -3,6 +3,7 @@ secrets:
   - "GH_SECRET_KEY=some_secret_key"
 environment:
   - "HAB_ORIGIN=edavis"
-  - "HAB_AUTH_TOKEN=asdf"
+  # - "HAB_AUTH_TOKEN=a_token"
 plan_path: "redis"
 repo_url: "https://github.com/elliott-davis/core-plans"
+channel: "unstable"
